streamlit/requirements.txt
streamlit==1.15.1
Pillow==9.3.0
loguru==0.6.0
opencv-python==4.6.0.66